Added score functionality and added some fields to classes. Going to clean up code.

+++ src/com/company/Bird.java
@@ -5,7 +5,7 @@ public class Bird {
     public double y;
     public double radius;
     public double yVel;
-    public double score;
+    public int score;
     public double gravity;
 
     Bird(double x, double y, double radius, double yVel) {
